feat: 통계 로직 수정

+++ src/main/java/com/smwu/donedone/done/ui/dto/StatisticsResponse.java
@@ -9,9 +9,9 @@ import lombok.Getter;
 public class StatisticsResponse {
     private Map<String, Integer> categoryCount;
     private Map<String, Integer> statusCount;
-    private Integer ratio;
+    private Double ratio;
 
-    public static StatisticsResponse of(Map<String, Integer> categoryCount, Map<String, Integer> statusCount, int ratio) {
+    public static StatisticsResponse of(Map<String, Integer> categoryCount, Map<String, Integer> statusCount, Double ratio) {
         return new StatisticsResponse(categoryCount, statusCount, ratio);
     }
 }
